+++ Source -compact mode/File2Startup/My Project/Settings.Designer.vb
@@ -78,6 +78,18 @@ Namespace My
                 Me("Culture") = value
             End Set
         End Property
+        
+        <Global.System.Configuration.UserScopedSettingAttribute(),  _
+         Global.System.Diagnostics.DebuggerNonUserCodeAttribute(),  _
+         Global.System.Configuration.DefaultSettingValueAttribute("True")>  _
+        Public Property CompactMode() As Boolean
+            Get
+                Return CType(Me("CompactMode"),Boolean)
+            End Get
+            Set
+                Me("CompactMode") = value
+            End Set
+        End Property
     End Class
 End Namespace